Make `catch` and `intercept` functions safe
This is a breaking change.

`throw` and every other throwing function is unsafe, so it's not
possible to cause UB with `catch`/`intercept` without other unsafe code.
And if `throw` is unsafe anyway, we can move the safety requirements to
the throwing site. This simplifies the API and lets us phrase docs
clearer. As a bonus, this reduces the size of `unsafe` code.